Update prerequisites-plaform-on-premises.doc
The order of the hardware requirements is O/S, System Data, Application Data, Docker, and etc; however, further down the document it explains each one (with the exception of O/S) in the order System Data, etc, Application Data, and then Docker.  The order should be the same as can cause confusion.